app/requirements.txt
graphene==3.1.0
fastapi==0.95.0
uvicorn==0.23.0
pyyaml==6.0
jinja2==3.1.2
python-multipart